+++ service/src/main/java/com/gmail/vpshulgaa/service/impl/OrderServiceImpl.java
@@ -4,7 +4,7 @@ import com.gmail.vpshulgaa.dao.OrderDao;
 import com.gmail.vpshulgaa.dao.entities.Item;
 import com.gmail.vpshulgaa.dao.entities.Order;
 import com.gmail.vpshulgaa.dao.entities.User;
-import com.gmail.vpshulgaa.dao.enums.Status;
+import com.gmail.vpshulgaa.dao.enums.StatusEnum;
 import com.gmail.vpshulgaa.service.ItemService;
 import com.gmail.vpshulgaa.service.OrderService;
 import com.gmail.vpshulgaa.service.UserService;
@@ -13,9 +13,12 @@ import com.gmail.vpshulgaa.service.converter.DtoConverter;
 import com.gmail.vpshulgaa.service.dto.ItemDto;
 import com.gmail.vpshulgaa.service.dto.OrderDto;
 import com.gmail.vpshulgaa.service.dto.UserProfileDto;
+import com.gmail.vpshulgaa.service.exception.EntityNotFoundException;
+import com.gmail.vpshulgaa.service.util.CurrentUserUtils;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -25,8 +28,8 @@ import org.springframework.transaction.annotation.Transactional;
 
 @Service
 public class OrderServiceImpl implements OrderService {
-    private static final Logger logger = LogManager.getLogger(OrderServiceImpl.class);
 
+    private static final Logger logger = LogManager.getLogger(OrderServiceImpl.class);
     private final OrderDao orderDao;
     private final Converter<OrderDto, Order> orderConverter;
     private final DtoConverter<OrderDto, Order> orderDtoConverter;
@@ -56,32 +59,25 @@ public class OrderServiceImpl implements OrderService {
     @Override
     @Transactional(readOnly = true)
     public OrderDto findOne(Long id) {
-        OrderDto orderDto = null;
-        try {
-            Order order = orderDao.findOne(id);
-            orderDto = orderDtoConverter.toDto(order);
-        } catch (Exception e) {
-            logger.error("Failed to get order", e);
+        Order order = orderDao.findOne(id);
+        if (order != null) {
+            return orderDtoConverter.toDto(order);
+        } else {
+            throw new EntityNotFoundException(Order.class, id);
         }
-        return orderDto;
-    }
-
-    @Override
-    @Transactional(readOnly = true)
-    public List<OrderDto> findAll() {
-        return new ArrayList<>();
     }
 
     @Override
     @Transactional
-    public OrderDto create(OrderDto orderDto, Long itemId, Long userId) {
+    public OrderDto create(OrderDto orderDto, Long itemId) {
         try {
             Order order = orderConverter.toEntity(orderDto);
+            Long userId = CurrentUserUtils.getPrincipal().getId();
             UserProfileDto userDto = userService.findOne(userId);
             ItemDto itemDto = itemService.findOne(itemId);
             order.setItem(itemConverter.toEntity(itemDto));
             order.setUser(userProfileConverter.toEntity(userDto));
-            order.setStatus(Status.NEW);
+            order.setStatus(StatusEnum.NEW);
             order.setCreated(LocalDateTime.now());
             orderDao.create(order);
             orderDto = orderDtoConverter.toDto(order);
@@ -124,27 +120,11 @@ public class OrderServiceImpl implements OrderService {
     @Override
     @Transactional
     public void deleteById(Long id) {
-        try {
+        if (orderDao.findOne(id) != null) {
             orderDao.deleteById(id);
-        } catch (Exception e) {
-            logger.error("Failed to delete order", e);
+        } else {
+            throw new EntityNotFoundException(Order.class, id);
         }
-    }
-
-    @Override
-    @Transactional(readOnly = true)
-    public List<OrderDto> findOrdersByUserId(Long userId) {
-        List<OrderDto> ordersDto = new ArrayList<>();
-        List<Order> orders;
-        try {
-            orders = orderDao.findordersByUserId(userId);
-            for (Order order : orders) {
-                ordersDto.add(orderDtoConverter.toDto(order));
-            }
-        } catch (Exception e) {
-            logger.error("Failed to find orders", e);
-        }
-        return ordersDto;
     }
 
     @Override
@@ -161,9 +141,10 @@ public class OrderServiceImpl implements OrderService {
 
     @Override
     @Transactional(readOnly = true)
-    public Long countOfOrderForUser(Long userId) {
+    public Long countOfOrderForUser() {
         Long count = 0L;
         try {
+            Long userId = CurrentUserUtils.getPrincipal().getId();
             count = orderDao.countOfOrdersForUser(userId);
         } catch (Exception e) {
             logger.error("Failed to find orders", e);
@@ -189,10 +170,11 @@ public class OrderServiceImpl implements OrderService {
 
     @Override
     @Transactional(readOnly = true)
-    public List<OrderDto> findOrdersByPageForUser(Long page, int maxResults, Long userId) {
+    public List<OrderDto> findOrdersByPageForUser(Long page, int maxResults) {
         List<OrderDto> ordersDto = new ArrayList<>();
         List<Order> orders;
         try {
+            Long userId = CurrentUserUtils.getPrincipal().getId();
             orders = orderDao.findOrdersByPageForUser(page, maxResults, userId);
             for (Order order : orders) {
                 ordersDto.add(orderDtoConverter.toDto(order));
@@ -201,5 +183,20 @@ public class OrderServiceImpl implements OrderService {
             logger.error("Failed to find orders", e);
         }
         return ordersDto;
+    }
+
+    @Override
+    @Transactional
+    public boolean isExistInOrders(Long itemId) {
+        boolean exists = false;
+        List<Order> orders = orderDao.findAll();
+        List<OrderDto> orderDtos = orderDtoConverter.toDtoList(orders);
+        for (OrderDto orderDto : orderDtos) {
+            if (Objects.equals(orderDto.getItemId(), itemId)) {
+                exists = true;
+                break;
+            }
+        }
+        return exists;
     }
 }
